+++ plugins/flytekit-async-fsspec/flytekitplugins/async_fsspec/__init__.py
@@ -11,6 +11,6 @@ This package contains things that are useful when extending Flytekit.
 """
 import fsspec
 
-from .s3fs import AsyncS3FileSystem
+from .s3fs.s3fs import AsyncS3FileSystem
 
 fsspec.register_implementation("s3", AsyncS3FileSystem)
